# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "FAQ Management System",
        "title": "FAQ API",
        "contact": {
            "name": "Ahmad Abdelrazik",
            "email": "ahmad.abdelrazik.swe@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/faq-categories/{category}/{id}/translations": {
            "get": {
                "description": "Retrieve every available language translation for a specific global FAQ",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Global FAQ Translations"
                ],
                "summary": "Get all translations for a global FAQ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category Name",
                        "name": "category",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "FAQ ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.getFaqTranslationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Category or FAQ not found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a new language translation to an existing global FAQ (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Global FAQ Translations"
                ],
                "summary": "Create global FAQ translation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category Name",
                        "name": "category",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "FAQ ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Translation Content",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.createTranslationInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.createTranslationResponse"
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Translation already exists",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/stores/{id}/faqs/{faq-id}/translations": {
            "get": {
                "description": "Retrieve every available language translation for a specific store's FAQ",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Store FAQ Translations"
                ],
                "summary": "Get all translations for a store FAQ",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Store ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "FAQ ID",
                        "name": "faq-id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.getFaqTranslationsResponse"
                        }
                    },
                    "404": {
                        "description": "Store or FAQ not found",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add a new language translation to an existing store FAQ (Merchant only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Store FAQ Translations"
                ],
                "summary": "Create store FAQ translation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Store ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "FAQ ID",
                        "name": "faq-id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Translation Content",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.createTranslationInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controllers.createTranslationResponse"
                        }
                    },
                    "409": {
                        "description": "Already exists or Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httputil.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.createTranslationInput": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
        "controllers.createTranslationResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "translation": {
                    "$ref": "#/definitions/controllers.translationDTO"
                }
            }
        },
...
        "controllers.getFaqTranslationsResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "default_language": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "translations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.translationDTO"
                    }
                }
            }
        },
...
        "controllers.translationDTO": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "faq_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
...
        "httputil.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        }
    }
}